refactor: LearningLogsPageからLogCardコンポーネントを分離
ログカードUI・CATEGORIES定数・getCategoryInfo/getCategoryColorを
LogCard.tsxに抽出。LearningLogsPage: 408→342行（66行削減）。
